piper-tts/requirements.txt
cffi==2.0.0
coloredlogs==15.0.1
flatbuffers==25.12.19
humanfriendly==10.0
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.23.2
packaging==26.0
piper-tts==1.4.1
protobuf==6.33.5
pycparser==3.0
soundfile==0.13.1
sympy==1.14.0
